requirements.txt
absl-py==2.0.0
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.42.0
blessed @ file:///croot/blessed_1701887495364/work
braceexpand==0.1.7
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
controlnet-aux==0.0.7
crcmod==1.7
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
Cython==3.0.7
dataclasses==0.6
datasets==2.13.0
diffusers==0.23.0
dill==0.3.6
dlib==19.24.2
docker-pycreds==0.4.0
easydict==1.11
edge-tts==6.1.9
einops==0.7.0
exceptiongroup==1.1.3
ExifRead-nocycle==3.0.1
fastapi==0.104.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock @ file:///croot/filelock_1672387128942/work
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
gpustat @ file:///croot/gpustat_1701888826609/work
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.6
-e git+https://huggingface.co/datasets/cosmicman/CosmicManHQ-1.0@52196b7a3ac9bb7d1227d9c6debe0cd95443d39b#egg=img2dataset&subdirectory=img2dataset
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
insightface==0.7.3
Jinja2 @ file:///croot/jinja2_1666908132255/work
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_core==5.5.0
jupytext==1.15.2
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
mediapipe==0.10.7
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mmcv-full==1.7.0
mmdet==2.26.0
mmengine==0.9.1
model-index==0.1.11
modelscope==1.9.4
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbformat==5.9.2
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
nvidia-ml-py @ file:///croot/nvidia-ml-py_1701875136273/work
onnx==1.15.0
onnxruntime==1.16.1
openai-clip==1.0.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.28
ordered-set==4.1.0
orjson==3.9.10
oss2==2.17.0
packaging==23.2
pandas==2.0.3
Pillow @ file:///croot/pillow_1696580024257/work
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
prettytable==3.9.0
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
pyarrow==14.0.0
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyre-extensions==0.0.29
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.4.3
python-dateutil==2.8.2
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.30.2
regex==2023.10.3
reportlab==4.0.8
requests==2.28.2
rich==13.4.2
rpds-py==0.12.0
ruff==0.1.15
safetensors==0.4.0
saliency==0.2.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.2
shellingham==1.5.4
simplejson==3.19.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
spaces==0.22.0
starlette==0.27.0
sympy @ file:///croot/sympy_1668202399572/work
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2023.7.10
tiktoken==0.5.2
timm==0.9.10
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.0
torchsummary==1.5.1
torchvision==0.15.0
tqdm==4.65.2
traitlets==5.13.0
transformers==4.31.0
transformers-stream-generator==0.0.4
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.24.0
wandb==0.16.1
wcwidth==0.2.12
webdataset==0.2.86
websockets==11.0.3
xformers==0.0.19
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
